Add initial support for querying remote repo lists.
This is currently little more than a demonstration of meaningful
communication between the plugin and Infocalypse containing the result
of a more complex operation.

After logging into a WoT identity and entering an identifier,
Infocalypse fetches and returns the repo list in usable form. There is
no progress indicator or automatic refreshing in the web UI, so the
current mode of operation is to check the console output for when
handling the query is complete and refresh the page. It's not currently
clear if periodic refreshes is a good way to improve this. Progress
messages would increase complication significantly.

Querying for the local repo list has been decreased in frequency, as it
was often enough that it seemed to crowd out the manual messages. Going
event-based would be very helpful here.

There are also some new TODOs about code design improvements for
handling query exchanges - including errors - and more concise page
registration.